backend/requirements.txt
fastapi
uvicorn
openai>=2.5.0
pydantic[email]
python-dotenv
sqlalchemy>=2.0
psycopg[binary]>=3.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==3.2.2
python-multipart
alembic>=1.13.0